blabla

+++ Jenkinsfile
@@ -28,11 +28,26 @@ pipeline {
                     exit 1
                 fi
 
-                # Supprimer les anciens ZIP dans /srv pour éviter l'encombrement
+                # Vérifier l'accès à /srv/
+                if [ ! -w /srv/ ]; then
+                    echo "❌ Erreur : Jenkins n'a pas les permissions en écriture sur /srv/"
+                    exit 1
+                fi
+
+                # Supprimer les anciens ZIP (si /srv/ accessible)
                 sudo find /srv/ -name 'Arkana_v*.zip' -type f -mtime +15 -delete
 
                 # Création de l'archive ZIP
                 zip -r ${ZIP_NAME} . -x "*.git*" -x "venv/*"
+
+                # Vérifier que le fichier ZIP a bien été créé
+                if [ ! -f ${ZIP_NAME} ]; then
+                    echo "❌ Erreur : L'archive ZIP n'a pas été créée correctement !"
+                    exit 1
+                fi
+
+                # Afficher la taille du fichier ZIP
+                ls -lh ${ZIP_NAME}
 
                 # Déplacer l'archive dans /srv/ avec gestion d'erreur
                 if sudo cp ${ZIP_NAME} /srv/; then
